Fix punishments not broadcasting on non-pgm setups
