+++ src/main/java/dev/pgm/community/moderation/feature/PGMPunishmentIntegration.java
@@ -1,15 +1,27 @@
 package dev.pgm.community.moderation.feature;
 
+import static net.kyori.adventure.text.Component.text;
+import static tc.oc.pgm.util.text.TemporalComponent.duration;
+
 import dev.pgm.community.Community;
 import dev.pgm.community.moderation.punishments.Punishment;
 import dev.pgm.community.moderation.punishments.types.MutePunishment;
+import java.time.Duration;
+import java.time.Instant;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.event.HoverEvent;
+import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
 import tc.oc.pgm.api.event.NameDecorationChangeEvent;
 import tc.oc.pgm.api.integration.Integration;
 import tc.oc.pgm.api.integration.PunishmentIntegration;
+import tc.oc.pgm.api.player.MatchPlayer;
+import tc.oc.pgm.events.PlayerParticipationStartEvent;
 
-public class PGMPunishmentIntegration implements PunishmentIntegration {
+public class PGMPunishmentIntegration implements PunishmentIntegration, Listener {
 
   private static final String BANNED_GROUP = "pgm.group.banned";
   private ModerationFeatureBase moderation;
@@ -20,6 +32,7 @@ public class PGMPunishmentIntegration implements PunishmentIntegration {
 
   public void enable() {
     Integration.setPunishmentIntegration(this);
+    Community.get().registerListener(this);
   }
 
   public void updateBanPrefix(Player player, boolean apply) {
@@ -45,5 +58,45 @@ public class PGMPunishmentIntegration implements PunishmentIntegration {
   @Override
   public boolean isHidden(Player player) {
     return moderation.getOnlineBan(player.getUniqueId()).isPresent();
+  }
+
+  @EventHandler
+  public void onMatchJoin(PlayerParticipationStartEvent event) {
+    MatchPlayer player = event.getPlayer();
+    if (moderation.getMatchBans() != null) {
+      Punishment punishment = moderation.getMatchBans().getIfPresent(player.getId());
+
+      if (punishment != null) {
+        Duration elasped = Duration.between(punishment.getTimeIssued(), Instant.now());
+        Duration remaining = moderation.getModerationConfig().getMatchBanDuration().minus(elasped);
+        Component reason =
+            text()
+                .append(text("You are banned from this match for "))
+                .append(duration(remaining, NamedTextColor.YELLOW))
+                .hoverEvent(
+                    HoverEvent.showText(
+                        text()
+                            .append(text("Reason: ", NamedTextColor.AQUA))
+                            .append(text(punishment.getReason(), NamedTextColor.GRAY))))
+                .build();
+        event.cancel(reason);
+      }
+    }
+
+    moderation
+        .getOnlineBan(player.getId())
+        .ifPresent(
+            ban -> {
+              Component reason =
+                  text()
+                      .append(text("You have been banned and are unable to play"))
+                      .hoverEvent(
+                          HoverEvent.showText(
+                              text()
+                                  .append(text("Reason: ", NamedTextColor.AQUA))
+                                  .append(text(ban.getReason(), NamedTextColor.GRAY))))
+                      .build();
+              event.cancel(reason);
+            });
   }
 }
